+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Allow all asynchronous operations to finish before exiting when running scripts.
 The fix that landed in version 1.4.1 still exhibited some issues ([#148](https://github.com/anmonteiro/lumo/issues/148)).
 - Fix installation on Windows ([#149](https://github.com/anmonteiro/lumo/issues/149)).
+- Don't crash if socket server can't bind to port ([#159](https://github.com/anmonteiro/lumo/issues/159)).
 
 ## [1.4.1](https://github.com/anmonteiro/lumo/compare/1.4.0...1.4.1) (2017-04-20)